feat: Store 뷰 페이지 구현
