--- src/app/store/components/review-list/review-list.component.html
@@ -1,3 +1,0 @@
-<p>
-  review-list works!
-</p>
