--- src/app/store/components/review-list/review-list.component.ts
@@ -1,14 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-review-list',
-  templateUrl: './review-list.component.html',
-  styleUrls: ['./review-list.component.scss'],
-})
-export class ReviewListComponent  implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {}
-
-}
