+++ src/app/store/components/store-info/store-info.component.html
@@ -1,3 +1,11 @@
-<p>
-  store-info works!
-</p>
+<!-- 가게 정보 -->
+<ion-content *ngIf="store">
+  <ion-card>
+    <ion-card-content>
+      <p><strong>카테고리:</strong> {{ store.category }}</p>
+      <p><strong>주소:</strong> {{ store.address }}</p>
+      <p><strong>전화번호:</strong> {{ store.contact_number }}</p>
+      <p *ngIf="store.description">{{ store.description }}</p>
+    </ion-card-content>
+  </ion-card>
+</ion-content>
