+++ src/app/store/components/store-info/store-info.component.ts
@@ -1,14 +1,12 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input } from '@angular/core';
+import { Store } from 'src/app/model/stores/read-store.interface';
 
 @Component({
   selector: 'app-store-info',
   templateUrl: './store-info.component.html',
   styleUrls: ['./store-info.component.scss'],
+  standalone: false
 })
-export class StoreInfoComponent  implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {}
-
+export class StoreInfoComponent {
+  @Input() store!: Store;
 }
