+++ src/app/store/pages/store-main/store.page.html
@@ -1,0 +1,11 @@
+<!-- Store 페이지 최상단 가게 이름 렌더링 -->
+<ion-header>
+  <ion-toolbar>
+    <ion-title *ngIf="store">{{ store.store_name }}</ion-title>
+  </ion-toolbar>
+</ion-header>
+
+<!-- store-info 렌더링 -->
+<ion-content>
+  <app-store-info *ngIf="store" [store]="store"></app-store-info>
+</ion-content>
