--- src/app/store/pages/store.page.ts
@@ -1,16 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-store',
-  templateUrl: './store.page.html',
-  styleUrls: ['./store.page.scss'],
-  standalone: false,
-})
-export class StorePage implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
