+++ src/app/store/store-routing.module.ts
@@ -1,11 +1,15 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { StorePage } from './pages/store.page';
+import { StorePage } from './pages/store-main/store.page';
 
 const routes: Routes = [
   {
     path: '',
+    component: StorePage
+  },
+  {
+    path: ':id',
     component: StorePage
   }
 ];
